Moved all current test classed to new packages
