+++ build.gradle
@@ -56,7 +56,7 @@ tasks.withType(JavaCompile) {
     options.encoding = 'UTF-8'
 }
 
-// Run pre tests for integeratin tests
+// Run pre tests for integration tests
 task preTests(type: Test) {
     useTestNG() {
         suites 'src/test/resources/integration_pre-tests.xml'
@@ -68,4 +68,15 @@ task integrationTests(type: Test) {
     useTestNG() {
         suites 'src/test/resources/integration_tests.xml'
     }
+
+    testLogging {
+        outputs.upToDateWhen { false }
+        showStandardStreams = true
+    }
+
+    options {
+        systemProperties(System.getProperties())
+    }
+
+    ignoreFailures = true
 }
