+++ src/test/java/eu/tsystems/mms/tic/testerra/plugins/xray/tests/jql/predefined/SummaryContainsExactTest.java
@@ -20,18 +20,21 @@
  *
  */
 
-package eu.tsystems.mms.tic.testerra.plugins.xray.jql.predefined;
+package eu.tsystems.mms.tic.testerra.plugins.xray.tests.jql.predefined;
 
 import static org.testng.Assert.assertEquals;
 
 
-import eu.tsystems.mms.tic.testerra.plugins.xray.AbstractTest;
+import eu.tsystems.mms.tic.testerra.plugins.xray.tests.AbstractTest;
+import eu.tsystems.mms.tic.testerra.plugins.xray.jql.predefined.SummaryContainsExact;
+import org.testng.Assert;
 import org.testng.annotations.Test;
 
-public class ProjectEqualsTest extends AbstractTest {
+public class SummaryContainsExactTest extends AbstractTest {
 
     @Test
     public void testCreateJql() throws Exception {
-        assertEquals(new ProjectEquals("BLA-1234").createJql(), "project = BLA-1234");
+        Assert.assertEquals(new SummaryContainsExact("bli bla blubb").createJql(), "summary ~ \"\\\"bli bla blubb\\\"\"");
+
     }
 }
