+++ src/test/java/eu/tsystems/mms/tic/testerra/plugins/xray/tests/jql/predefined/TestTypeEqualsTest.java
@@ -20,20 +20,21 @@
  *
  */
 
-package eu.tsystems.mms.tic.testerra.plugins.xray.jql;
+package eu.tsystems.mms.tic.testerra.plugins.xray.tests.jql.predefined;
 
 import static org.testng.Assert.assertEquals;
 
 
-import eu.tsystems.mms.tic.testerra.plugins.xray.AbstractTest;
-import java.util.Arrays;
+import eu.tsystems.mms.tic.testerra.plugins.xray.tests.AbstractTest;
+import eu.tsystems.mms.tic.testerra.plugins.xray.jql.predefined.TestType;
+import eu.tsystems.mms.tic.testerra.plugins.xray.jql.predefined.TestTypeEquals;
+import org.testng.Assert;
 import org.testng.annotations.Test;
 
-public class MultiValueTest extends AbstractTest {
+public class TestTypeEqualsTest extends AbstractTest {
 
     @Test
-    public void testToString() {
-        final MultiValue multiValue = new MultiValue(Arrays.asList("bli", "bla", "blubb"));
-        assertEquals(multiValue.createJql(), "(bli, bla, blubb)");
+    public void testCreateJql() throws Exception {
+        Assert.assertEquals(new TestTypeEquals(TestType.AutomatedGeneric).createJql(), "\"Testtyp\" = \"Automated[Generic]\"");
     }
 }
