+++ src/test/java/eu/tsystems/mms/tic/testerra/plugins/xray/testundertest/annotation/MethodsAndClassAnnotatedTest.java
@@ -24,7 +24,7 @@ package eu.tsystems.mms.tic.testerra.plugins.xray.testundertest.annotation;
 
 import eu.tsystems.mms.tic.testerra.plugins.xray.annotation.XrayTest;
 import eu.tsystems.mms.tic.testerra.plugins.xray.annotation.XrayTestSet;
-import eu.tsystems.mms.tic.testerra.plugins.xray.testundertest.AbstractTestBase;
+import eu.tsystems.mms.tic.testerra.plugins.xray.pretests.demotests.AbstractTestBase;
 import org.testng.annotations.Test;
 
 @XrayTestSet(key = "SWFTE-7")
