+++ src/test/resources/testsets/_connector_tests.xml
@@ -4,7 +4,7 @@
 
     <test name="Connector" parallel="methods" thread-count="8">
         <packages>
-            <package name="eu.tsystems.mms.tic.testerra.plugins.xray.connect"/>
+            <package name="eu.tsystems.mms.tic.testerra.plugins.xray.tests.connect"/>
         </packages>
     </test>
 </suite>
