+++ src/test/resources/testsets/_jql_tests.xml
@@ -4,8 +4,8 @@
 
     <test name="JQL" parallel="methods" thread-count="8">
         <packages>
-            <package name="eu.tsystems.mms.tic.testerra.plugins.xray.jql"/>
-            <package name="eu.tsystems.mms.tic.testerra.plugins.xray.jql.predefined"/>
+            <package name="eu.tsystems.mms.tic.testerra.plugins.xray.tests.jql"/>
+            <package name="eu.tsystems.mms.tic.testerra.plugins.xray.tests.jql.predefined"/>
         </packages>
     </test>
 </suite>
